requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.3
six==1.17.0
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0